src/requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
binance==0.3
blinker==1.8.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.1.1
cryptography==42.0.7
cycler==0.12.1
dateparser==1.1.8
et-xmlfile==1.1.0
filelock==3.14.0
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.4.0
future==1.0.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h2o==3.46.0.1
h5py==3.10.0
html5lib==1.1
idna==3.4
imbalanced-learn==0.12.2
instaloader==4.11
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.3.2
keras==3.3.3
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.41.1
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.4
multitasking==0.0.11
namex==0.0.8
nltk==3.8.1
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
peewee==3.17.0
Pillow==10.1.0
praw==7.7.1
prawcore==2.4.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.19.0
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
PySocks==1.7.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scikeras==0.13.0
scikit-learn==1.4.2
scipy==1.11.3
seaborn==0.13.0
selenium==4.21.0
shap==0.43.0
six==1.16.0
slicer==0.0.7
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.5
ta==0.11.0
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.14.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.18.0.post0
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
tzdata==2023.3
tzlocal==5.1
ujson==5.8.0
update-checker==0.18.0
urllib3==2.0.7
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.31
